refactor: centralize business error mapping
- unwrap nested BusinessException in ExceptionResponseMapper
- reuse mapper results in bad request handling
- validate and normalize datasource type on save and update
