+++ data-agent-backend/src/main/java/io/github/malonetalk/exception/GlobalExceptionHandler.java
@@ -68,9 +68,9 @@ public class GlobalExceptionHandler {
         HttpMessageNotReadableException.class
     })
     public ResponseEntity<Result<Object>> handleBadRequest(Exception exception) {
-        BusinessException businessException = findBusinessException(exception);
-        if (businessException != null) {
-            return response(exceptionResponseMapper.resolve(businessException));
+        ErrorResponse errorResponse = exceptionResponseMapper.resolve(exception);
+        if (errorResponse.errorCode() != ErrorCode.INTERNAL_ERROR) {
+            return response(errorResponse);
         }
         return response(ErrorCode.BAD_REQUEST, resolveBadRequestMessage(exception));
     }
@@ -215,17 +215,6 @@ public class GlobalExceptionHandler {
             return propertyPath;
         }
         return propertyPath.substring(separatorIndex + 1);
-    }
-
-    private BusinessException findBusinessException(Throwable exception) {
-        Throwable current = exception;
-        while (current != null) {
-            if (current instanceof BusinessException businessException) {
-                return businessException;
-            }
-            current = current.getCause();
-        }
-        return null;
     }
 
     private ResponseEntity<Result<Object>> response(ErrorCode errorCode) {
